+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to DBC Utility will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-11-10
+
+### Changed
+- Changed to UV package manager
+- Added button to create a new DBC file
+- Added buttons to reorder messages and signals
+- Binded the Edit message and Edit signals to double click
+- Added Buttons to duplicate Message and duplicate signals
+
+### Fixed
+- Version label at the bottom is using the real version (from pyproject.toml)
+
 ## [1.0.1] - 2025-01-29
 
 ### Changed